+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Optimize execution unit setup time.
 - Vectorized *Get Spline Length* node.
 - Avoid unnecessary updates caused by *Object Visibility Output* node.
+- Avoid unnecessary updates caused by *Viewer* node.
 
 
 ## 2.2.2 (16 August 2021)